Phase-II-Todo-Full-Stack-Web-Application/todo-web-app/src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Gemini Agent <gemini@google.com>"]
readme = "README.md"
packages = [{include = "backend", from = "src"}]

[project]
name = "backend"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.31",
    "alembic>=1.13.1",
    "uvicorn>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "bcrypt>=5.0.0",
    "psycopg2-binary>=2.9"
]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
sqlmodel = "^0.0.31"
alembic = "^1.13.1"
uvicorn = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.2.1"
python-multipart = "^0.0.21"
bcrypt = "^5.0.0"
psycopg2-binary = "^2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (<0.24.0)",
]
